REPLACE mocks to folder mocks/
